refactor(contextual_backend/transport): extract frame io
Create a trait to handle the frame reading logic and pass this to the
server on startup. This create a better separation of concerns by
introducing layers in the server. We still need to improve the
ergonomics for the server creation. Defining `F` each time is a pain in
the ass...